More initial steps, new square conflict plan
This isn't ready to compile yet and I made the changes outside eclipse as
indicated by uncommented pseudocode in the initializer. All that remains
are the file inputs for the initializer (clues and square) and the con-
flict checks.

Checking the row and column are simple enough. The square, as before,
gave me pause. Originally, I intended to read a grid file and determine
which square I was in and which part of the square, to get to the upper
left corner. Then simply check rows and columns within a square's bounds.

But, the same effect is achieved by using a shorter file that shows the
square boundaries of that grid type. Puzzles with non-rectangular areas
are no longer sudoku puzzles, and thus out of scope. Frankly, I'll pare
the square file to just the puzzle configuration information in the next
commit. Corrupted configurations are a concern but outside my interest
presently. Frankly, if I'm to include it, best to just stick it all in
the puzzle file.

As far as checking a square, the focus will be less than some multiple of
the square boundaries, so the offset is found and applied.