Stack overflow
My recursion logic is almost certainly flawed. When I changed to a puzzle
of 4x4, it produced several hundred or thousand calls. Only 64 are
required for a blank grid, so it must be my fault.

To start, I will iterate through a cell's internal guess rather than using
recursion. If the calls are still unreasonable, then I may branch to a
totally iterative algorithm.