Still overflowing, on 4x4 even
Something deep is wrong in the flow. This generates hundreds of calls
when 64 should be the upper limit. I've corrected some bugs but this
still eludes me. (The conflicts seem to check out now.) Frankly, I'm
pretty close to really studying someone else's code to compare mine
against. (Also inclined to not make these hard line breaks in the commit.)