Solved some overflow, eerie terminations
It turns out that when I backtracked from an untenable guess onto a clue,
the clue always pushed the focus forward in an infinite loop. If I had
considered the calls more carefully, I would have been in a better
position to realize that. So, for the moment, there's a backtracking flag
until I think of something nicer.

Unfortunately, it seems I may have fouled up that implementation anyway.
The 4x4 puzzle reported failure when it had half the puzzle solved.
Further, the 9x9 overflowed trying to backtrack past 0,0. I
need to vet thisFirstCell(), in case it isn't preventing the backtrack.